legacy: batchuploader removal
* INCOMPATIBLE Removes legacy batchuploader module. (addresses #3233)
