+++ invenio/ext/legacy/layout.py
@@ -202,7 +202,6 @@ class WebInterfaceInvenio(WebInterfaceDirectory):
         'help',
         'exporter',
         'kb',
-        'batchuploader',
         'bibsword',
         'ping',
         'admin2',
@@ -234,9 +233,7 @@ class WebInterfaceInvenio(WebInterfaceDirectory):
         exporter='WebInterfaceFieldExporterPages',
         kb='WebInterfaceBibKnowledgePages',
         admin2='WebInterfaceAdminPages',
-        batchuploader='WebInterfaceBatchUploaderPages',
         bibsword='WebInterfaceSword',
-        ping='WebInterfacePingPages',
         textmining='WebInterfaceDocExtract',
         yourcomments='WebInterfaceYourCommentsPages',
         goto='WebInterfaceGotoPages',
