+++ invenio/legacy/websession/webaccount.py
@@ -90,7 +90,7 @@ def perform_youradminactivities(user_info, ln):
             your_admin_activities.append(action)
 
     if SUPERADMINROLE in your_roles:
-        for action in ("runbibedit", "cfgbibformat", "cfgoaiharvest", "cfgoairepository", "cfgbibrank", "cfgbibindex", "cfgwebaccess", "cfgwebcomment", "cfgwebsearch", "cfgwebsubmiit", "cfgbibknowledge", "runbatchuploader"):
+        for action in ("runbibedit", "cfgbibformat", "cfgoaiharvest", "cfgoairepository", "cfgbibrank", "cfgbibindex", "cfgwebaccess", "cfgwebcomment", "cfgwebsearch", "cfgwebsubmiit", "cfgbibknowledge", ):
             if action not in your_admin_activities:
                 your_admin_activities.append(action)
