+++ setup.py
@@ -270,7 +270,6 @@ setup(
             'plotextractor = invenio.utils.scripts.plotextractor:main',
             # Legacy
             'alertengine = invenio.legacy.webalert.scripts.alertengine:main',
-            'batchuploader = invenio.legacy.bibupload.scripts.batchuploader',
             'bibcheck = invenio.legacy.bibcheck.scripts.bibcheck:main',
             'bibcircd = invenio.legacy.bibcirculation.scripts.bibcircd:main',
             'bibauthorid = '
